Overhaul memory access and index typing
This commit restructures memory access to exclusively use the
shared-mutable memory access unified by the `radium` crate, which
prevents construction of any bare references which may introduce UB
when contending for memory.

In addition, bit index patterns are encoded in the type system,
which permits removal of some runtime validity checks and
strengthens awareness of the proper arithmetic.

The `BitSlice` type still needs work done to remove production of
contended references, but the main portion of the work is done.